email/llm/requirements.txt
# OpenAI LLM Service Requirements

# Core OpenAI API
openai>=1.3.0

# FastAPI and dependencies
fastapi>=0.104.0
uvicorn>=0.24.0

# Data validation and serialization
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client
httpx>=0.25.0

# Logging
structlog>=23.0.0

# Optional: For async support
aiofiles>=23.0.0
